pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-news-digest"
version = "0.1.0"
description = "AI-powered personalized news digest for Telegram channels. Learns your writing style."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "feedparser>=6.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.scripts]
news-digest = "news_digest.cli:main"

[project.optional-dependencies]
dev = ["ruff", "pytest"]
